Group send modal specific components together
