+++ src/modals/SendModals/SendModalDeployContract.tsx
@@ -20,11 +20,8 @@ import { convertAlphToSet } from '@alephium/sdk'
 import { binToHex, contractIdFromAddress, SignDeployContractTxResult } from 'alephium-web3'
 import { useState } from 'react'
 
-import GasSettingsExpandableSection from '../../components/GasSettingsExpandableSection'
 import InfoBox from '../../components/InfoBox'
-import AddressSelectFrom from '../../components/Inputs/AddressSelectFrom'
 import AmountInput from '../../components/Inputs/AmountInput'
-import BytecodeInput from '../../components/Inputs/BytecodeInput'
 import Input from '../../components/Inputs/Input'
 import { useAddressesContext } from '../../contexts/addresses'
 import { Client } from '../../contexts/global'
@@ -33,8 +30,11 @@ import useDappTxData from '../../hooks/useDappTxData'
 import useStateObject from '../../hooks/useStateObject'
 import { DeployContractTxData, TxPreparation } from '../../types/transactions'
 import { hasNoGasErrors, isAmountWithinRange } from '../../utils/transactions'
+import AddressSelectFrom from './AddressSelectFrom'
 import AlphAmountInfoBox from './AlphAmountInfoBox'
 import BuildTxFooterButtons from './BuildTxFooterButtons'
+import BytecodeInput from './BytecodeInput'
+import GasSettingsExpandableSection from './GasSettingsExpandableSection'
 import { ModalInputFields } from './ModalInputFields'
 import SendModal, { TxContext } from './SendModal'
 import { CheckTxProps, PartialTxData } from './types'
