+++ src/modals/SendModals/SendModalScript.tsx
@@ -19,11 +19,8 @@ along with the library. If not, see <http://www.gnu.org/licenses/>.
 import { convertAlphToSet } from '@alephium/sdk'
 import { SignExecuteScriptTxResult } from 'alephium-web3'
 
-import GasSettingsExpandableSection from '../../components/GasSettingsExpandableSection'
 import InfoBox from '../../components/InfoBox'
-import AddressSelectFrom from '../../components/Inputs/AddressSelectFrom'
 import AmountInput from '../../components/Inputs/AmountInput'
-import BytecodeInput from '../../components/Inputs/BytecodeInput'
 import { useAddressesContext } from '../../contexts/addresses'
 import { Client } from '../../contexts/global'
 import { useSendModalContext } from '../../contexts/sendModal'
@@ -31,8 +28,11 @@ import useDappTxData from '../../hooks/useDappTxData'
 import useStateObject from '../../hooks/useStateObject'
 import { ScriptTxData, TxPreparation } from '../../types/transactions'
 import { hasNoGasErrors, isAmountWithinRange } from '../../utils/transactions'
+import AddressSelectFrom from './AddressSelectFrom'
 import AlphAmountInfoBox from './AlphAmountInfoBox'
 import BuildTxFooterButtons from './BuildTxFooterButtons'
+import BytecodeInput from './BytecodeInput'
+import GasSettingsExpandableSection from './GasSettingsExpandableSection'
 import { ModalInputFields } from './ModalInputFields'
 import SendModal, { TxContext } from './SendModal'
 import { CheckTxProps, PartialTxData } from './types'
