+++ src/modals/SendModals/SendModalTransfer.tsx
@@ -20,9 +20,7 @@ import { convertAlphToSet } from '@alephium/sdk'
 import { SignTransferTxResult } from 'alephium-web3'
 import { useCallback, useState } from 'react'
 
-import GasSettingsExpandableSection from '../../components/GasSettingsExpandableSection'
 import InfoBox from '../../components/InfoBox'
-import AddressSelectFrom from '../../components/Inputs/AddressSelectFrom'
 import AmountInput from '../../components/Inputs/AmountInput'
 import Input from '../../components/Inputs/Input'
 import { useAddressesContext } from '../../contexts/addresses'
@@ -33,8 +31,10 @@ import useStateObject from '../../hooks/useStateObject'
 import { TransferTxData, TxPreparation } from '../../types/transactions'
 import { isAddressValid } from '../../utils/addresses'
 import { hasNoGasErrors, isAmountWithinRange } from '../../utils/transactions'
+import AddressSelectFrom from './AddressSelectFrom'
 import AlphAmountInfoBox from './AlphAmountInfoBox'
 import BuildTxFooterButtons from './BuildTxFooterButtons'
+import GasSettingsExpandableSection from './GasSettingsExpandableSection'
 import { ModalInputFields } from './ModalInputFields'
 import SendModal, { TxContext } from './SendModal'
 import { CheckTxProps, PartialTxData } from './types'
